+++ CHANGES.rst
@@ -4,6 +4,24 @@ Change Log
 4.0 (unreleased)
 ----------------
 
+- Add property to clear session data at login boundary to the session auth
+  helper. This property defaults to ``False`` to preserve the current behavior.
+  Clearing session data during login helps mitigate session fixation attacks:
+  https://owasp.org/www-community/attacks/Session_fixation
+
+- Fire ``UserSessionStarted`` event when a new cookie-based session starts.
+
+- Fire ``UserSessionStarted`` event when a new user session starts.
+
+- Fire ``UserLoggedOut`` event if the credentials are reset in the user folder.
+
+- Add ``UserLoggedIn`` event that is firing upon successful user login.
+
+- Add ``UserLoggedOut`` event that is firing upon successful user logout.
+
+- Add ``UserSessionStarted`` event that is firing when a new user session is
+  created.
+
 - Add support for Python 3.13.
 
 - Drop support for Python 3.8.